only add readlimit connectors to external connectors ()
if you use session auth, the driver will pipeline the `LOGOFF`, `LOGON`,
and `RUN` messages,
This meant the read limit connector will be added between the logoff and
logon.
if your parameter map in the run is large (like in SPD) this will hit
the read limit and fail the query.
meaning using session auth with a large parameter map will throw, even
if you would be authenticated properly, we can fix
this for SPD and query router by not installing it at all for internal
trusted connections.