Skip TransactionLogChecker multi-version-in-one-file-check in benign case ()
For TX logs, there is an edge case in which tx pull pulls everything up
to and including the upgrade transaction and then starts the store and
fails to rotate the tx log. This is benign though for non-segmented logs
and won't be fixed. In order to minimize false positives, we skip
throwing here if we find that there is a checkpoint exactly at the tx id
of the upgrade transaction. The benign case described will do recovery
before starting the DB, and thus it will always have a checkpoint with
the tx id of the upgrade tx.

**Cherry-pick this to 5.26**

---------

Co-authored-by: fabiim <dev.fbotelho@gmail.com>