Add rewriter for fully-fused MERGE ()
Introduce a rewriter that checks ahead-of-time if we can fully fuse the
MERGE clause and if so rewrites it to a new plan FusedMerge. Since the
rewriting happens before the actual code-generation it needs to be on
the conservative side so that we don't get any false positives.

Combined with the recently added support of non-fused MERGE (still
behind a feature flag) this allows us to have some parts of the plan
fused while the MERGE part of the query is not.