Keep old Cypher Shell behaviour on old Bolt versions. ()
If the driver is newer then the server it might polyfill GqlStatus
object. We do not want to expose those to the users.

This was changed as part of
https://github.com/neo-technology/neo4j/pull/32076 and I think we did
not consider the case when Cypher Shell is run toward an old server (in
the flaky test 4.3) closely enough

Custom builds to verify that this works
4.3:
https://live.neo4j-build.io/buildConfiguration/BuildServicePipelines_neo4jv43cyphertest/25654607
dev:
https://live.neo4j-build.io/buildConfiguration/BuildServicePipelines_neo4jcyphertest/25654627