Enable non-fused merges in SPD by default ()
- Enables non-fused MERGE in SPD transactions
- Update RemoteBatchingStrategy to plan batching for queries containing
non-locking MERGE

Non-fused MERGE in pipelined runtime is still hidden behind a feature
toggle. However, we want to make it enabled by default in SPD
transactions by default to begin with it since this is the main driver
for this work. The non-fused implementation currently doesn't support
locking merge, i.e. merging on a relationship where any of nodes are
bound such as `MATCH (from), (to) MERGE (from)-[r:R]->(to)`. Hence, we
still need an exception for patterns like that RemoteBatchingStrategy.