Add property constraints to DynamicLabelNodeLookup ()
We will begin to support index seeks in the dynamic node scan operator,
and also rename it to `DynamicLabelNodeLookup`.

This PR:
✅ Adds support for property filtering of existing scan operators in all
runtimes
✅ Adds support for index seek lookup in Slotted and Interpreted in very
particular circumstances
❌ Does not add any index seek lookups in Pipelined or Parallel

Once this PR is merged, Planner can provide exact property predicates to
this plan and be assured that all runtimes will perform a property
filter on each found node. In the meantime, Runtime can continue adding
support for different index circumstances

---------

Co-authored-by: Rob Steward <speaktorob@gmail.com>