Delete include-cypher profile (-DskipCypher)
IntelliJ doesn't handle the combination of profile activated sub-modules
and transitive dependencies correctly.
See:
* [How to perform "reload all Maven projects" (non-incrementally) with
the classic
UI?](https://intellij-support.jetbrains.com/hc/en-us/community/posts/21354698847250-How-to-perform-reload-all-Maven-projects-non-incrementally-with-the-classic-UI?page=1#community_comment_21547745402642)
* [Profiles activated via Maven Tool Window do not apply to Maven Run
Configurations](https://youtrack.jetbrains.com/issue/IDEA-359439/Profiles-activated-via-Maven-Tool-Window-do-not-apply-to-Maven-Run-Configurations)
* [{maven profile} | 2k+
issues](https://youtrack.jetbrains.com/issues/IDEA?q=%7Bmaven%20profile%7D)

I boiled it down to a [minimal maven
project](https://github.com/neo-technology/maven-profile-sync-issue).
It's just 4 modules, a bunch of java classes, nothing more. I can get it
to fail systematically when switching between branches. The presence of
a profile and of transitive depedencies seem necessary for it to fail.
Note that it is timing sensitive, you need to trigger the build at the
right time after changing branch. This seems consistent with what we see
in the monorepo.

This might also be behind the "No Scala SDK in module" error as Scala is
a dependency of the Cypher modules. IntelliJ loads the Scala SDK based
on what it finds in the maven dependencies.

This is the only profile used to exclude sub-modules in a POM where some
of the sub-modules are used as transitive dependencies elsewhere. There
is `neo4j-uring` which also excludes sub-modules but there is no
dependencies as of now. Also it's based on the JDK version probably
leading to a full rebuild in any case.

All the other profiles like `skip-cypher-it` are used to exclude test
suites. I think they're fine – famous last words.

I believe this will help, but it won't fix everything. I also suspect
merging this will upset the IntelliJ caches in some instances. It will
probably take a version bump or two to stabilise. Let's see how we do
when we get to 2025.10.0-SNAPSHOT. In the meantime, "Reload All Maven
Projects" is your friend.