fix(allReduce): ensure all variables are imported when using call(*) within ()
The behaviour with a CALL(*) within an allReduce was different from a
CALL(<importVariable>).
This PR ensures that when declaring the reductionStep variable in the
nested scope, all the variables from the parent scope are included as
well.
This could lead to duplicate variable declaration errors for which we
included an "override" flag.
This PR adds a few tests to check if the declared variable in the
reduction step can shadow a groupVariable.

This PR also cleans up the previously introduced "groupVariable" flag
from symbols, since it is not used anymore.