Rewrite GBPTree value updates for MVCC compatibility ()
In MVCC, we don't want to kill the old key-value when setting the value
to something else, as it would lose the history. So the logic in
`InternalTreeLogic.setValueAtWithFallback` has to be changed to not kill
the old key-value. And since `InternalTreeLogic` is agnostic to whether
we're versioned or not, the non-versioned part had to be changed to work
with the new behaviour in there. Any killing of old values is now
handled in the leaf itself, where it's also inserted.

And also since we're not killing before a potential split anymore,
finding the split position had to be adjusted to account for that. More
specifically, we need to calculate the split by imagining that the old
space is gone and the new space is there.

The versioned part will be changed in another PR.

---------

Co-authored-by: Lukas Gustavsson <lukas.gustavsson@neo4j.com>