TriadicSelection support for SPD ()
TriadicSelection matches on specific subplans in the logical plan.
Matching on `Selection` causes a problem for SPD plan.
Finding selections is a bit more complicated in SPD, since it can be
represented by a combination of `Selection`, `RemoteBatchProperties` and
`RemoteBatchPropertiesWithFilter`.
This PR adds an unapply `SpdSelections` which matches the different
combinations to obtain all selections.
It is used in `triadicSelectionFinder` such that SPD plans can be
translated into triadicSelection plans.

Additionally, this PR adds an internal database configuration
`internal.cypher.planning_selector_candidates_maximum` that is used in
`Selector.scala` instead of the hardcoded value `3`.
`RemoteBatchPropertiesTriadicSelectionPlanningIntegrationTest` shows an
example of the impact of increasing it to `4` on the generated plan.