Inject through externalId into DatabaseLog ()
Aura uses JSON logging and for multidatabase they would like to tag the
log lines by every database. This is already present, but the problem is
that the two added tags (name and database id) are not really sufficient
for Aura, as they don't use either of those to identify the database.
They use their own database id (external id) which is different from the
(internal) database id.

With the use of the databases reconciler the external id set in the
topology graph and is accessible inside Neo4j. After it can be
even sent in via the `CREATE DATABASE` command.

This change extends the `DatabaseLog*` classes to allow the usage of it.
The reason why the external id is just chucked alongside the
`NamedDatabaseId` is that in case we want to introduce even more
identifiers in the future (as the external id in fact could be already
build up from two parts), the `DatabaseLogIdentifier` can encapsulate
it.