fix(cost model): index seek should be cheaper than index scan with filter ()
Ensure that even when the selectivity isn't great, on SPD
```
.nodeIndexSeek(predicate)
```
is cheaper than
```
.filter(predicate)
.nodeIndexScan
```
by adding the cost of using cacheProperties in filters and also a cost
to actually fetching properties from the index, i.e., property ->
GetValue

As a side effect, composite indexes that only need to fetch one property
are cheaper than composite indexes that need to fetch multiple
properties.
As a related fix, we also ensure that we don't get values for properties
for composite indexes when we only want to check if they are not null.

There is an accompanying microbenchmark which locally showed that
.nodeIndexSeek is a little quicker than a scan with filter.

Also, enabled a bunch of acceptance tests.