Make range index do equality comparisons on point/duration types edge case ()
This ensures that when dealing with the types point/duration, and array
variants of these two
on an range index we will still match on queries in the form `y <= x <=
y` when `y==x`.

This is "relevant" because right now it is the behaviour we have when
not using an index,
and we should strive for consistency. No one has ever complained about
this, though, and the
behaviour of comparison operators on these types is certainly not well
defined or consistent,
so users shouldn't expect much from them (hopefully).

First and foremost, I'm doing this to address some of the SPD cypher
acceptance tests.
In SPD, a query in the form `x >= y` won't find any match when `x == y`
for the mentioned types
over a range index. This happens because the shard query that ends up
going to the shards
is in the form of `y <= x <= y` and that would never have any results
before this PR.

cl : Ensure range index is consistent with no index when doing equality
comparisons over points, durations and arrays variants of the two.