Disable flaky test under enveloped log profile ()
Disable the
TransactionLogFileRotateAndReadRaceIT.shouldNotSeeEmptyLogFileWhenReadingTransactionStream
test when on enveloped log files, because we cannot guarantee the legacy
behaviour when active reads and writes overlap. In actual usage e.g.
CDC, or the merged RAFT/tx logs we will have strict barriers by
position, or appendIndex to prevent encountering the unstable active
ends of files.