feat(AllReduce): Error Messages ()
Adds a new AllReduce function to catch arguments written in the format
`allReduce(expr1,... exprN)` without the correct reduce expression using
the | in the syntax. 
It ignores the function for Cypher5
Adds a semantic test catching other syntax errors where a reduce
expression with | is used but other arguments are missing.