New envelope streaming protocol ()
This introduces messages and handlers for dealing with enveloped logs.
It's not tailored to a particular form of enveloped log (i.e. should
work with both the raft log, tx log and in the future, the merged one).
**Note** that the "protocol" is not intended to be used in production.
The only reason there's a protocol is for testing purposes. The idea is
to use the handlers and messages in other protocols, for example CATCHUP
and RAFT.

------

The gist of it all is: 
The client asks for a range of entries (`LogRangeRequest`). If the
client has existing entries, it must send the previousChecksum in the
request. This is to make sure that the logs are the same.

The server, receives the request in `LogRangeRequestHandler` and
validates the indexes. Thereafter, it creates a `ChunkedSegmentSender`.
This component is responsible for sending the segments containing the
requested entries, chunked up in a configurable size. The sender in turn
makes use of `SegmentCursor`. This cursor opens a `EnvelopeReadChannel`
and returns the segments one by one, as long as there are segment within
the entry boundaries.

Each segment contains the segment payload and some metadata: 
1. Size (the size of the segment + metadata) 
2. Last (is this segment the last one the server will send?)
3. PrevTerm (since the envelope header doesn't contain the prev term, it
has to be sent separately.)
4. Start offset (the absolute offset, i.e. where in the log does this
segment start)

When the client receives the segments, they might be chunked up, so it
puts all received chunks in a `InboundSegmentCursor`. This one returns
the segments, only when they're complete. The consumer of the segments
is the `EnvelopedLogWriter`. This class keeps track of the local state
of the log. It knows if a new log has to be created, if the current one
needs to rotate etc. And it's also responsible for taking the incoming
segments and writing them down to the local files. Apart from this, it
also has logic for making sure that the offsets are aligned.

------

Two FYIs:

<del>1. There are some discussions around using the
`EnvelopeWriteChannel.directPutAll()` instead of `EnvelopedLogWriter`.
The benefit is that [the padding
problem](https://github.com/neo-technology/neo4j/pull/32075/commits/7b86964d7a501701530a9596508e3c361e39cd67)
is fixed in that method. It makes sure that the client adds the padding
if it didn't receive it from the server. It hasn't yet been decided what
to use. </del> It's now fixed in this PR.

2. This
[bug](https://github.com/neo-technology/neo4j/pull/32075/commits/79de62f10b4f6809dc50b44f64231e7d44d94efb)
is being fixed in an ongoing PR. So feel free to ignore this commit. It
just has to be there to not break things.

---------

Co-authored-by: Grant Lodge <6323995+thelonelyvulpes@users.noreply.github.com>