fix(SPD): ShardOperatorPushdownStrategy fetch variables and predicates ()
Ensure that only the alreadyCachedRewriter also looks at the property
instead of the variable alone when rewriting a predicate used in
RemoteBatchPropertiesWithFilter. The predicate should not use
cacheProperty instead of property for its predicates. This fixes
NodeIndexContainsScanAcceptanceTest where the winning plan was cheaper
because the filter used cached properties.

Also, while fetching properties in remoteBatchPropertiesWithFilter,
filter out properties that have already been fetched.