feat(GSO): endpoint constraints for undirected and disjunction of types ()
Implements graph schema optimisations for undirected
relationship patterns and when relationship patterns have several
relationship types (in disjunction).

This information is used in 3 places:
1. In a leaf planner to plan an index
2. For skipping label filters
3. For cardinality estimation

To implement this, this PR
- unifies the code paths for 1 and 3 to both use same method
`impliedEndpointLabelsMap`.
- splits this functionality into parts:
1. iterate relationships in `impliedEndpointLabelsMap` (split useful for
cardinality estimation where we have one relationship at hand)
2. iterate types in `internalImplicationsFromRelationship` (split needed
for relationship type disjunction)
  3. collect implications in `implicationsFromRelationshipType`
- moves filtering on whether a label was already explicitly present
outside to where it is needed (node index leaf planner)

Also, the PR
- Introduces caching on `internalImplicationsFromRelationship`
- introduces `enableGraphSchemaOptimisation()` into logical plan builder
to conveniently turn GSO on and off.
- Improves Cardinality Estimation: If `-[:R]-` implies `:B`, then we
should not assume `:B` to be selective at all when estimating
cardinality of a relationship of type `:R`: `card(()-[:R]->(:A:B)) =
card(()-[:R]->(:A))`