feat(errors): add a parser rule to catch bad allReduce syntax ()
The expected signature for allReduce is : `allReduce(accumulator ::
VARIABLE = initial :: ANY, variable :: VARIABLE IN list :: LIST<ANY> |
expression :: ANY, predicate :: BOOLEAN ) :: BOOLEAN.`
When an expression without `|` is passed in, it is parsed as a
FunctionInvocation, and SemanticAnalysis will throw a meaningful error
message.
However, if there is an expression with the pipe operator, but we are
missing some other arguments, the syntax errors are no longer useful.
For example,
1. `allReduce( a = 5, a IN A | a + 1)` throws a syntax error `"Invalid
input ')': expected an expression or ','"` while attempting to parse to
an allReduce expression
2. `allReduce(acc =0, acc + 5, iter IN r | acc + 1)` throws a syntax
error `Invalid input '|': expected an expression, ')' or ','"` while
attempting to parse to a FunctionInvocation

To avoid confusion, we add a catch all allReduceInvalidArguments rule that
would catch expressions that do not match allReduce expressions and
throw a standard invalid arguments error

Changelog: The GQL error code on an invalid allReduce invocation is now
42I51