Fix bug with nested ordered unions in pipelined runtime ()
After a recent change in the logic of dynamic scheduling, the
OrderedUnionOperator now needs to make sure it has explicitly requested
input from both lhs and rhs, before it can proceed to output the
remaining rows in the bubble-phase of scheduling.

The scheduling phase needs to be set back to Tunnel before visiting the
other branch after visiting a pipeline in the Bubble phase.

This fixes a bug where a nested ordered union could output remaining
rows too early, which could lead to rows arriving out of order.