feat(Cardinality Estimation): Use label info from `:A|B` ()
Label info represents information on which node we **know** to have
which label. This information is then used to get more exact cardinality
estimation, for example from indexes.

If in the query, there is a predicate `:A|B`, neither of these two
labels is for certain. Therefore, we couldn't infer any label info, not
being able to utilise the information that indexes could provide.

This change tries to make that possible by identifying such predicate
and estimating each option (`:A` and `:B`) separately.

In order to be able to combine these again, we remove the disjunction
`:A|B` from the predicates, calculate that cardinality and identify how
much adding each label reduces it (thus the selectivity of this label). We
can then combine the selectivities and apply the combined selectivity to
the base cardinality to get an improved overall cardinality estimation.
(This is done in
`AssumeIndependenceQueryGraphCardinalityModel.getBaseQueryGraphCardinalityWithInferredLabelContext`.)

---------

Co-authored-by: Sachin Vasant <sachin.vasant@neo4j.com>