Make block command type an enum ()
`BlockCommand` type should be an enum.

We introduce a `public enum BlockCommand.BlockCommandType` with values
for each of the existing block commands, explicitly constructed with a
`byte type` which has the command value. This introduces a single source
of truth for the byte value of block commands.

* The enum has a `COMMAND_TYPE_ARRAY` from types to enum values.
* We delete the `static type` fields of all `BlockCommand` subclasses,
and refer to `BlockCommandType.<subtype>.type` where we need the `byte`
type of a subcommand (i.e. on serialization)
* Reading a command now consists of mapping the `byte` to an enum using
the map, and switching on the enum.

We implement a new collection helper, `ByteToEnum`, as suggested by
@klaren, with which we implement the `COMMAND_TYPE_ARRAY` of
`BlockCommand`. `ByteToEnum` uses a fixed size (256 entry) array of
enums, i.e. an entry per each possible byte value. We expect most of
these will be `null`, but the space penalty does not seem excessive; we
are trying ensure no lost performance.

---------

Co-authored-by: Matthew Parnell <matt@parnmatt.co.uk>