Allow dbms.recreateDatabase to be used on SPD ()
`dbms.recreateDatabase` is related to a physical database, SPD is built
up from several physical databases. It is possible to just recreate
shards one-by-one already, but unless a partial disaster it is not
really useful to do it one-by-one.
In case of recreate from different backup it is completely fragile and
tedious to use.

Now `dbms.recreateDatabase` can recreate a whole SPD (all shards) in one
go both for disaster recovery and for recreate from different backup.

For latter the same `seedUri` format must be used as in `CREATE
DATABASE`. For disaster recovery either specify `seedingServers:[]` for
system selection or `seedingServers:{foo-g000:['a','b'], ...}` for
defining specific server seeding.