Fix an issue with shortestPath in fused pipelined runtime ()
Because of how code generation works we cannot reuse the generated
intermediate representation multiple times. This lead to an issue where
we weren't properly checking so that the target node was satisfying the
provided predicate. This can lead to errors for queries using

```cypher
MATCH p=shortestPath(...)
WHERE all(n IN nodes(p) WHERE [some predicate])
```


changelog: Fix bug that sometimes lead to the wrong result for queries
with `p=shortestPath()` combined with `all(n IN nodes(p) WHERE...)'