Change SubtractionNodeByLabelsScan CostPerRow to match other scans ()
With this change we now get ties for `MATCH (n:A&!B)` for
SubtractionNodeByLabelsScan vs NodeByLabelScan+Filter. This is handled
by adding a case in LeafPlanSelectorHeuristic. With additional labels
SubtractionNodeByLabelsScan becomes cheaper.

The change in calculating positive labels cardinality was needed to
handle `MATCH (n:A&B&!C)` case where we compare
SubtractionNodeByLabelsScan vs IntersectionNodeByLabelsScan+Filter. The
cost of Filter only applies to the incoming rows after the intersection,
i.e. "positive labels cardinality", which made it cheaper than the
original SubtractionNodeByLabelsScan formula.