Make failure to register built-in procedures a critical error ()
All kinds of features of the DBMS depend on the fact that the built-in
procedures are available. This PR makes this implicit fact explicit by
refusing to start the DBMS if the build-in procedures are not loaded
correctly (this is really a programming error).

As a side benefit, the PR moves the awkward registration of the
`SpecialBuiltInProcedures` into the ordinary flow.