Fix non-parallel recovery for vector stores ()
This fixes some issues found with recovery with vector stores, though
the effect is likely not seen until they are switched on in another PR.

* If the vector store files are not there, it still might be recoverable
with the transaction logs if the reason for their absence is a crash
after creation before checkpointing.

* If there are no transaction logs to recover from, we should check
again, as those files are truly missing, and we are missing data.

* After reverse recovery, we should be in a good state; therefore, if we
are still missing files known to that state, then those files truly are
missing, and we are missing data.

* Introduced VectorValueReader{,s} and VectorValueWriter{,s} with
appropriately named get methods so it is clear you are specifically
getting a reader or a writer.

* Introduced VectorStoreHolderMutator, which specifically handles the
creating and removing of the VectorStore instances. The vector stores
needed to be removed from the internal cache, and closed idempotently,
as life would try and close it on shutdown.

* Both feature enabled and disabled paths need to be able to operate on
the stores in recovery. Restructing most operations under an abstract
vector store holder, and specialising behaviour, allows things to be
reasoned together.

* Switched approach to create different vector store holders from a
factory to be passed around. This allowed for Recovery to use a lenient
writer, which doesn't care about the existence of a store; important for
backwards-recovery. Normal paths do care, and should through if a store
is missing.

* We were trying to use a transactional vector store creation during
recovery; with the new recovery vector store holders and restrictive
writer interface, this is mitigated.

* A few classes and methods are also renamed to minimise overloading
terms. It was found difficult to discuss and reason about when it was
ambiguous which stage in the process was being addressed.