Add support for dynamic sized values in versioned GBPTree ()
This PR adds support for dynamic sized values in versioned block
GBPTree. First off the `LeafNodeDynamicSizeVersioned` got changed. And
we got rid of the padding. It works similarly to the non versioned
`LeadNodeDynamicSize`. The difference is in the format, and the fact
that we have to save the old value before killing it. Once it's time for
the new value to be written, we push the old value to the history. Which
brings us to the new `DynamicSizeTreeValueHistory`. The old
`TreeValueHistory`, which is now called `FixedSizeTreeValueHistory` only
supported fixed value sizes. But since there are still use cases for
this, I kept it, and just added another implementation for the dynamic
one.