Get rid of test-jar deps on neo4j-ast ()
Test-jar dependencies lead to all kind of issues.
In particular it leads to maven occasionally retieving an outdated
previously built test-jar leading to class not found errors and the
like.
There's no convenient way of having tests depend on the tests of another
module.
In this PR I split `neo4j-ast` into three modules:
* `neo4j-ast`: only keeping the source folder
* `neo4j-test-util`: test utilities for `neo4j-ast` including
`AstConstructionTestSupport`
* `neo4j-ast-test`: tests for `neo4j-ast`, depends on the two other
module

This change is part of: [Make it possible to run `./mvnw
compile`](https://trello.com/c/UyLZEJWn/2-make-it-possible-to-run-mvnw-compile)