Remove deprecated usage of DatabaseShutdownException ()
This usage is a bit strange. The prefetcher is owned by the global
module. If it is stopped, the whole server is shutting down, but we
don't have a good message for that. The database should already be
offline at this point, so I think it is effectively dead code