Fix bug in unique constraint creation that could lead to corrupt unique indexes. ()
Exact seek problem
------------------

When populating a unique index, we add duplicates anyway (but record
them), and re-verify the duplicates later.
`See BlockBasedIndexPopulator.handleMergeConflict`

E.g. let's say we observe node(id=42){prop: "foo"}, and then later
node(id=100){prop: "foo"}.
At that point, our tree might look like this:

```
Level 0
{3,leaf,generation=3,keyCount=4} [String("aaa")],entityId=0=[no value][D] [String("foo")],entityId=42=[no value][D] [String("foo")],entityId=100=[no value][D] [String("zzz")],entityId=12=[no value][D]
```

We proceed to add more data, leading to a split in pages. After all
values are added, the tree might look like this:

```
Level 0
{6,internal,generation=3,keyCount=1}  /3\ [String("foo")],entityId=100 /5\
Level 1
{3,leaf,generation=3,keyCount=4} [String("aaa")],entityId=0=[no value][D]   [String("bbb")],entityId=1=[no value][D]  [String("ccc")],entityId=2=[no value][D]  [String("foo")],entityId=42=[no value][D]
{5,leaf,generation=3,keyCount=4} [String("foo")],entityId=100=[no value][D] [String("xxx")],entityId=10=[no value][D] [String("yyy")],entityId=11=[no value][D] [String("zzz")],entityId=12=[no value][D]
```

This is a correct split: While adding things to three, we first try to
add them "with uniqueness" (`NativeIndexKey.compareId=false`).
Only if that fails we add them "without uniqueness"
(`NativeIndexKey.compareId=true`).
Regardless, if adding data leads to a split, it will simply copy the key
that is immediately right of the split
to the new parent to become the splitter key. In our case
[String("foo")],entityId=100.
This is correct, as the GBPTree always assumes keys to be unique.

Later, when verifying uniqueness for all the recorded duplicates (
BlockBasedIndexPopulator.verifyUniqueKeys ),
we traverse the tree with an exact seek to search for
[String("foo")],entityId=42 OR String("foo")],entityId=100.
The outcome is identical, it doesn't matter which of the 2 is recorded
as the offending one.

We do this exact seek "with uniqueness"
(NativeIndexKey.compareId=false). So when traversing down the tree and
comparing splitter keys against our search key, we will not compare
entity IDs.
This leads to us comparing our search key against the splitter key
[String("foo")],entityId=100,
which will always be `0`. That means go RIGHT.
We then traverse to leaf 5, and find only 1 key with String("foo"), not
2.

The outcome is that we can succeed creating a uniqueness constraint in
the presense of non-unique data.
2 scenarios need to be distinguished:

1) The DB crashes before we can make a checkpoint: When doing recovery,
we replay the log and
populate the index again. The physical layout might very well have
changed, so that this time around
we do not succeed and the index that backs our constraint never comes
online.
Trying to commit any transaction that updates the index will throw a
`UnableToValidateConstraintException`
   because the index is in a failed state.
The constraint can be dropped, data cleaned up and then a new attempt at
creating the constraint can be made.
2) The constraint and index are successfully checkpointed: At this point
we have a corrupt DB.
It will, however, most likely never be noticed, since operation can
resume normally, and the index will,
after the initial population, not allow any more duplicates to be
created.

Backwards seek problem
----------------------

This is probably more of a theoretical problem, since we don't do
backwards search on unique indexes while they are
in this weird limbo state where duplicates are actually allowed.

This problem occurs even when the duplicates are not spread out over
multiples pages. Let's consider this tree again:

```
Level 0
{3,leaf,generation=3,keyCount=4} [String("aaa")],entityId=0=[no value][D] [String("foo")],entityId=42=[no value][D] [String("foo")],entityId=100=[no value][D] [String("zzz")],entityId=12=[no value][D]
```

When we now do a backwards seek from String("foo"), "with uniqueness"
(`NativeIndexKey.compareId=false`), to the beginning of the index, the
following happens:
We do a binary search inside the leaf to find the position of
String("foo").
See `KeySearch.search`. This binary search is written in such a way that
it will always find the leftmost position,
if there are duplicates.

Therefore, our seek will return

```
[String("foo")],entityId=42=[no value][D]
[String("aaa")],entityId=0=[no value][D]
```
and miss [String("foo")],entityId=100=[no value][D] as the first entry
that should have been returned.

Solution
--------

Both problems are addressed by forbidding seeks when
`NativeIndexKey.compareId=false`.
Instead, the entity IDs should be set to `Long.MIN_VALUE` and
`Long.MAX_VALUE` respectively.

---

cl:
Fix a bug where, in very rare cases, a uniqueness constraint would
mistakenly allow duplicate properties during population of the
constraint backing index.